+++ packages/dill-docs/astro.config.mjs
@@ -19,8 +19,11 @@ export default defineConfig({
 	output: "server",
 	// adapter: deno(),
 	adapter: netlify({
-		// edgeMiddleware: true
+		// edgeMiddleware: true,
+		// cacheOnDemandPages: true,
+		imageCDN: false,
 	}),
+	site: "https://dill.tylerbutler.com",
 	integrations: [
 		starlight({
 			title: "dill",