+++ packages/repopo/src/index.ts
@@ -1,29 +1,27 @@
 export { run } from "@oclif/core";
-
-export type {
-	PolicyFailure,
-	PolicyFixResult,
-	PolicyFunctionArguments,
-	PolicyHandler,
-	PolicyInstance,
-	PolicyInstanceSettings,
-	PolicyHandlerResult,
-	PolicyName,
-	PolicyStandaloneResolver,
-	PolicyDefinition,
-} from "./policy.js";
-
 export type {
 	// PolicyList,
 	RepopoConfig,
 } from "./config.js";
 export { makePolicy } from "./makePolicy.js";
-export {
-	definePackagePolicy as generatePackagePolicy,
-	type PackageJsonHandler,
-} from "./policyDefiners/definePackagePolicy.js";
+export type {
+	PolicyDefinition,
+	PolicyFailure,
+	PolicyFixResult,
+	PolicyFunctionArguments,
+	PolicyHandler,
+	PolicyHandlerResult,
+	PolicyInstance,
+	PolicyInstanceSettings,
+	PolicyName,
+	PolicyStandaloneResolver,
+} from "./policy.js";
 export {
 	defineFileHeaderPolicy as generateFileHeaderPolicy,
 	type FileHeaderGeneratorConfig,
 	type FileHeaderPolicyConfig,
 } from "./policyDefiners/defineFileHeaderPolicy.js";
+export {
+	definePackagePolicy as generatePackagePolicy,
+	type PackageJsonHandler,
+} from "./policyDefiners/definePackagePolicy.js";